moved edit button below
